third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/injected-inline-script-allowed-expected.txt

ALERT: Pass 1 of 2
ALERT: Pass 2 of 2
